gnn_service/requirements.txt
fastapi
uvicorn[standard]
neo4j
mysql-connector-python
python-dotenv
torch
torch-geometric
pandas
numpy
matplotlib